Create a new source file for pimCmdFuse.
